+++ feedback/default.nix
@@ -7,7 +7,7 @@
 }:
 mkDerivation {
   pname = "feedback";
-  version = "0.1.0.5";
+  version = "0.2.0.0";
   src = ./.;
   isLibrary = true;
   isExecutable = true;